move push helper commands into util
