+++ src/commands/tx.js
@@ -1,6 +1,3 @@
-const path = require('path');
-const chalk = require('chalk');
-
 module.exports = {
 	command: 'tx',
 	description: 'transifex lifecycle integration commands',
